Hometask #4 - second attempt.
